requirements.txt
django
mysqlclient
django-redis
django-ipware
python-decouple
graphene-django
django-user-agents